+++ tests/TestCase/Command/ExportCommandTest.php
@@ -17,16 +17,19 @@ namespace DatabaseBackup\Test\TestCase\Command;
 
 use Cake\Console\ConsoleIo;
 use Cake\Console\Exception\StopException;
+use Cake\Console\TestSuite\ConsoleIntegrationTestTrait;
 use Cake\Console\TestSuite\StubConsoleOutput;
 use DatabaseBackup\Command\ExportCommand;
-use DatabaseBackup\TestSuite\CommandTestCase;
+use DatabaseBackup\TestSuite\TestCase;
 use DatabaseBackup\Utility\BackupExport;
 
 /**
  * ExportCommandTest class
  */
-class ExportCommandTest extends CommandTestCase
+class ExportCommandTest extends TestCase
 {
+    use ConsoleIntegrationTestTrait;
+
     /**
      * @var string
      */
@@ -101,7 +104,7 @@ class ExportCommandTest extends CommandTestCase
         $this->exec($this->command . ' --rotate 3 -v');
         $this->assertExitSuccess();
         $this->assertOutputRegExp('/Backup `[\w\-\/\:\\\\]+backup_[\w_]+\.sql` has been exported/');
-        $this->assertOutputContains('Backup `' . basename(array_value_first($files)) . '` has been deleted');
+        $this->assertOutputContains('Backup `' . basename($files[0]) . '` has been deleted');
         $this->assertOutputContains('<success>Deleted backup files: 1</success>');
         $this->assertErrorEmpty();
     }