+++ tests/TestCase/Command/RotateCommandTest.php
@@ -15,13 +15,16 @@ declare(strict_types=1);
  */
 namespace DatabaseBackup\Test\TestCase\Command;
 
-use DatabaseBackup\TestSuite\CommandTestCase;
+use Cake\Console\TestSuite\ConsoleIntegrationTestTrait;
+use DatabaseBackup\TestSuite\TestCase;
 
 /**
  * RotateCommandTest class
  */
-class RotateCommandTest extends CommandTestCase
+class RotateCommandTest extends TestCase
 {
+    use ConsoleIntegrationTestTrait;
+
     /**
      * @test
      * @uses \DatabaseBackup\Command\RotateCommand::execute()